(finally) moved shremote_new.py -> shremote.py
